+++ lib/ovto.rb
@@ -9,6 +9,8 @@ if RUBY_ENGINE == 'opal'
   require_relative 'ovto/state'
   require_relative 'ovto/version'
   require_relative 'ovto/wired_actions'
+  require_relative 'ovto/wired_action_set'
+  require_relative 'ovto/middleware'
 else
   require 'ovto/version'
   require 'opal'; Opal.append_path(__dir__)
@@ -55,5 +57,28 @@ module Ovto
       }
     end
     raise ex
+  end
+
+  # Something like `obj.meth(state: state, **args, &block)`
+  # Safe even if `obj.meth` does not have `state:`
+  def self.send_args_with_state(obj, meth, args, state, &block)
+    parameters = obj.method(meth).parameters
+    accepts_state = `!parameters` || parameters.nil? || parameters.any?{|item|
+      item == [:key, :state] ||
+      item == [:keyreq, :state] ||
+      item[0] == :keyrest
+    }
+    if accepts_state
+      # We can pass `state:` safely
+      args_with_state = {state: state}.merge(args)
+      return obj.__send__(meth, args_with_state, &block)
+    else
+      # Check it is empty (see https://github.com/opal/opal/issues/1872)
+      if args.empty?
+        return obj.__send__(meth, &block)
+      else
+        return obj.__send__(meth, **args, &block)
+      end
+    end
   end
 end